64-bit fix: don't cram pointers into Special longs

+++ src/game/Player_Control.c
@@ -756,7 +756,7 @@ ObjNode	*newObj,*newObj2;
 	MakeObjectTransparent(newObj, .5);							// make transparent
 	MakeObjectKeepBackfaces(newObj);							// also keep backfaces
 	
-	theNode->JetFlameObj = (UInt32)newObj;						// point to this flame
+	theNode->JetFlameObj = newObj;								// point to this flame
 	theNode->ExhaustTimer = 0;									// init exhaust timer
 
 		/* MAKE LEFT FLAME */	
@@ -769,8 +769,8 @@ ObjNode	*newObj,*newObj2;
 	
 	MakeObjectTransparent(newObj2, .5);							// make transparent
 	
-	newObj->JetFlameObj = (UInt32)newObj2;						// other flame points to this flame	
-	newObj2->JetFlameObj = (UInt32)theNode;						// this flame points back to player
+	newObj->JetFlameObj = newObj2;								// other flame points to this flame	
+	newObj2->JetFlameObj = theNode;								// this flame points back to player
 	
 			/* JET SOUND */
